requirements.txt
cassandra-driver
fastapi
gunicorn
jupyter
pandas
numpy
pypyr
python-dotenv
tensorflow
uvicorn
keras
boto3
pypyr
python-dotenv
joblib
lightgbm
scikit-learn
pydantic-settings